Allow multiple ConInfo objects.
This allows us to keep the previous convention's data around when it comes time to start a new convention,
instead of having to clear it each year. This removes the restriction that there is only one ConInfo object
at a time, and adds the get_current_con function to utils.py, which returns the current convention. The
current logic is just that the newest ConInfo object is the current con, but since that logic is isolated
to one utility function it'll be easy to change that out if we come up with a better way in the future.

Right now, there is no way to see any data from previous cons except by going through the raw database
data on the admin page. We can add that later if need be.